% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{ipca_vm}
\alias{ipca_vm}
\title{IPCA with vertex expansion}
\usage{
ipca_vm(interval_data, id_col = NULL, dims = 2)
}
\arguments{
\item{interval_data}{A data frame with paired interval columns.}

\item{id_col}{Identifier or grouping column.}

\item{dims}{Number of reduced dimensions.}
}
\value{
A reduced-space interval data frame.
}
\description{
IPCA with vertex expansion
}
