% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{itsne-package}
\alias{itsne}
\alias{itsne-package}
\title{Interval t-SNE for interval-valued data}
\description{
The \code{itsne} package provides interval-valued extensions of t-SNE, interval
baseline methods, plotting helpers, modified-LCMC utilities, and reproducible
workflows for the manuscript's simulation and real-data studies.
}
\author{
\strong{Maintainer}: Han-Ming Wu \email{hanmingwu1103@gmail.com}

}
\keyword{internal}
