% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_interval_method_grid}
\alias{plot_interval_method_grid}
\title{Plot a multi-method interval comparison grid}
\usage{
plot_interval_method_grid(
  methods,
  titles = names(methods),
  ncol = 4,
  shared_limits = FALSE,
  pad = 0.05,
  legend_position = "bottom"
)
}
\arguments{
\item{methods}{A named list of reduced-space interval data frames.}

\item{titles}{Optional character vector of panel titles.}

\item{ncol}{Number of columns.}

\item{shared_limits}{Whether to use a common square plotting region.}

\item{pad}{Plot-padding multiplier.}

\item{legend_position}{Legend position.}
}
\value{
A patchwork object.
}
\description{
Plot a multi-method interval comparison grid
}
