% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselines.R
\name{ipca_qm}
\alias{ipca_qm}
\title{IPCA with quantile expansion}
\usage{
ipca_qm(interval_data, id_col = NULL, m = 4, dims = 2)
}
\arguments{
\item{interval_data}{A data frame with paired interval columns.}

\item{id_col}{Identifier or grouping column.}

\item{m}{Number of quantile segments.}

\item{dims}{Number of reduced dimensions.}
}
\value{
A reduced-space interval data frame.
}
\description{
IPCA with quantile expansion
}
