% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itsne-package.R
\docType{package}
\name{itsne-package}
\alias{itsne}
\alias{itsne-package}
\title{Interval t-SNE for interval-valued data}
\description{
The \code{itsne} package provides interval-valued extensions of t-SNE, interval
baseline methods, plotting helpers, modified-LCMC utilities, and reproducible
workflows for the manuscript's simulation and real-data studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hanmingwu1103/I-tSNE}
  \item Report bugs at \url{https://github.com/hanmingwu1103/I-tSNE/issues}
}

}
\author{
\strong{Maintainer}: Han-Ming Wu \email{hanmingwu1103@gmail.com}

Authors:
\itemize{
  \item Han-Ming Wu \email{hanmingwu1103@gmail.com}
}

}
\keyword{internal}
