% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmc.R
\name{plot_lcmc_grid}
\alias{plot_lcmc_grid}
\title{Plot modified-LCMC curves}
\usage{
plot_lcmc_grid(lcmc_tables, method_name_map = NULL, palette = NULL, nrow = 1)
}
\arguments{
\item{lcmc_tables}{Output of compute_lcmc_tables().}

\item{method_name_map}{Optional named character vector for legend labels.}

\item{palette}{Optional named vector of colors.}

\item{nrow}{Number of facet rows.}
}
\value{
A ggplot object.
}
\description{
Plot modified-LCMC curves
}
